requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
jinja2==3.1.6
httpx==0.28.1
gitpython==3.1.45
streamlit>=1.44.0,<2.0.0
plotly>=6.0.0,<7.0.0
pandas>=2.2.0,<3.0.0
